Give BITS_IN_SET type int
This type seems to be assumed in other parts of the code.

+++ config/common_part.h
@@ -113,7 +113,7 @@
 /* Number of bits in a Modula-2 set.  Modula-2 set is always implemented in C 
    by unsigned int. */
 
-#define BITS_IN_SET (CHAR_BIT*sizeof(unsigned int))		/*unsigned int*/
+#define BITS_IN_SET (int)(CHAR_BIT*sizeof(unsigned int))		/*unsigned int*/
 
 /* Byte is synonym of char. */
